Allow local urls
